feat: remove 'symlinks' option
The preStart scripts are enough: just make a script symlinking
what you need, if you even really need it. Hermes does not, as
for now, follow symlinks: that means I will just generate the
pages directly to the web hosting directory, as there is no real
reason to not do that.

Signed-off-by: Francesco Saccone <francesco@francescosaccone.com>